Create newReddit.sol
This is supposed to be the start of Reddit style forums based on Ethereum staking. Contributers are encouraged to stake value into communities. 

Incidentally, a new idea could be to have a form of debate such that it always required three parties. In a debate, two sides would form a lock of which the value could be derived by the third party in the unwinding and arbitration of an issue.